+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+* Added support for offline use through the `AUSPICE_ENABLE_SERVICE_WORKER` environment variable read at build-time. This is added primarily for auspice.us, but should work for any Auspice server where assets are served at the root path.
+  ([#2062](https://github.com/nextstrain/auspice/pull/2062))
+
 ## version 2.72.0 - 2026/06/30